+++ axelor-stock/src/main/java/com/axelor/apps/stock/web/InventoryLineController.java
@@ -17,6 +17,7 @@
  */
 package com.axelor.apps.stock.web;
 
+import com.axelor.apps.stock.db.Inventory;
 import com.axelor.apps.stock.db.InventoryLine;
 import com.axelor.apps.stock.service.InventoryLineService;
 import com.axelor.exception.AxelorException;
@@ -34,7 +35,11 @@ public class InventoryLineController {
       throws AxelorException {
 
     InventoryLine inventoryLine = request.getContext().asType(InventoryLine.class);
-    inventoryLine = inventoryLineService.updateInventoryLine(inventoryLine);
+    Inventory inventory =
+        request.getContext().getParent() != null
+            ? request.getContext().getParent().asType(Inventory.class)
+            : inventoryLine.getInventory();
+    inventoryLine = inventoryLineService.updateInventoryLine(inventoryLine, inventory);
     response.setValue("rack", inventoryLine.getRack());
     response.setValue("currentQty", inventoryLine.getCurrentQty());
   }